backend/requirements.txt
# Dependencies for FastAPI Backend
fastapi>=0.85.0
uvicorn>=0.18.0
python-multipart>=0.0.5
pillow>=9.0.0
numpy>=1.21.0
tensorflow-cpu>=2.8.0 # Using CPU version for lighter weight serving
python-dotenv>=0.21.0
